requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Mail==0.9.1
Flask-WTF==1.1.1
Werkzeug==2.3.7
email-validator==2.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
blinker==1.6.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
